% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R, R/plots.R
\name{cv.exclusive_lasso}
\alias{cv.exclusive_lasso}
\alias{plot.ExclusiveLassoFit_cv}
\title{CV for the Exclusive Lasso}
\usage{
cv.exclusive_lasso(
  X,
  y,
  groups,
  ...,
  family = c("gaussian", "binomial", "poisson"),
  offset = rep(0, NROW(X)),
  weights = rep(1, NROW(X)),
  type.measure = c("mse", "deviance", "class", "auc", "mae"),
  nfolds = 10,
  parallel = FALSE
)

\method{plot}{ExclusiveLassoFit_cv}(x, bar.width = 0.01, ...)
}
\arguments{
\item{X}{The matrix of predictors (\eqn{X \in \R^{n \times p}}{X})}

\item{y}{The response vector (\eqn{y})}

\item{groups}{An integer vector of length \eqn{p} indicating group membership.
(Cf. the \code{index} argument of \code{\link[grplasso]{grpreg}})}

\item{...}{Additional arguments passed to \code{\link{exclusive_lasso}}.}

\item{family}{The GLM response type. (Cf. the \code{family} argument of
\code{\link[stats]{glm}})}

\item{offset}{A vector of length \eqn{n} included in the linear
predictor.}

\item{weights}{Weights applied to individual
observations. If not supplied, all observations will be equally
weighted. Will be re-scaled to sum to \eqn{n} if
necessary. (Cf. the \code{weight} argument of
\code{\link[stats]{lm}})}

\item{type.measure}{The loss function to be used for cross-validation.}

\item{nfolds}{The number of folds (\eqn{K}) to be used for K-fold CV}

\item{parallel}{Should CV run in parallel? If a parallel back-end for the
\code{foreach} package is registered, it will be used. See the
\code{foreach} documentation for details of different backends.}

\item{x}{An \code{ExclusiveLassoFit_cv} object as produced by
\code{\link{cv.exclusive_lasso}}.}

\item{bar.width}{Width of error bars}
}
\description{
CV for the Exclusive Lasso
}
\details{
As discussed in Appendix F of Campbell and Allen [1], cross-validation
   can be quite unstable for exclusive lasso problems. Model selection by BIC
   or EBIC tends to perform better in practice.
}
\examples{
n <- 200
p <- 500
groups <- rep(1:10, times=50)
beta <- numeric(p);
beta[1:10] <- 3

X <- matrix(rnorm(n * p), ncol=p)
y <- X \%*\% beta + rnorm(n)

exfit_cv <- cv.exclusive_lasso(X, y, groups, nfolds=5)
print(exfit_cv)
plot(exfit_cv)

# coef() and predict() work just like
# corresponding methods for exclusive_lasso()
# but can also specify lambda="lambda.min" or "lambda.1se"
coef(exfit_cv, lambda="lambda.1se")
}
\references{
Campbell, Frederick and Genevera I. Allen. "Within Group Variable Selection
    with the Exclusive Lasso." Electronic Journal of Statistics 11(2),
    pp.4220-4257. 2017. \doi{10.1214/EJS-1317}
}
