% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.ExclusiveLassoFit}
\alias{plot.ExclusiveLassoFit}
\title{Visualizing Exclusive Lasso Regularization Paths}
\usage{
\method{plot}{ExclusiveLassoFit}(x, xvar = c("norm", "lambda", "l1"),
  label = FALSE, legend = "topright", ...)
}
\arguments{
\item{x}{An \code{ExclusiveLassoFit} object produced by \code{\link{exclusive_lasso}}}

\item{xvar}{Value to use on the x-axis (ordinate). \itemize{
\item \code{norm}: The value of the composite \eqn{\ell_2/\ell_1}{l2/l1} norm
    "exclusive lasso" norm
\item \code{lambda}: The log of the regularization parameter \eqn{\lambda}{lambda}.
\item \code{l1}: The \eqn{\ell_1}{l1}-norm of the solution
}}

\item{label}{Should variables be labeled? Set \code{label=NULL} to label only
those variables which are non-zero at the sparsest end of
the regularization path.}

\item{legend}{Should a legend of top variables (heuristically selcted) be displayed?
Set \code{legend=NULL} to disable, else give the location of the legend.}

\item{...}{Additional arguments passed to plotting functions
n <- 200
p <- 500
groups <- rep(1:10, times=50)
beta <- numeric(p);
beta[1:10] <- 3

X <- matrix(rnorm(n * p), ncol=p)
y <- X %*% beta + rnorm(n)

exfit <- exclusive_lasso(X, y, groups)
plot(exfit)
plot(exfit, legend=NULL, xvar="lambda")}
}
\description{
Visualizing Exclusive Lasso Regularization Paths
}
