% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExclusiveLasso.R
\name{coef.ExclusiveLassoFit}
\alias{coef.ExclusiveLassoFit}
\alias{predict.ExclusiveLassoFit}
\title{Make predictions using the exclusive lasso.}
\usage{
\method{coef}{ExclusiveLassoFit}(object, lambda = s, s = NULL, exact = FALSE, group_threshold = FALSE, ...)

\method{predict}{ExclusiveLassoFit}(
  object,
  newx,
  lambda = s,
  s = NULL,
  type = c("link", "response", "coefficients"),
  group_threshold = FALSE,
  exact = FALSE,
  offset,
  ...
)
}
\arguments{
\item{object}{An \code{ExclusiveLassoFit} object produced by \code{\link{exclusive_lasso}}.}

\item{lambda}{The value of the regularization paramter (\eqn{lambda}) at which to
return the fitted coefficients or predicted values. If not supplied, results for
the entire regularization path are returned. Can be a vector.}

\item{s}{An alternate argument that may be used to supply \code{lambda}. Included for
compatability with \code{\link[glmnet]{glmnet}}.}

\item{exact}{Should the exclusive lasso be re-run for provided values of \code{lambda}?
If \code{FALSE}, approximate values obtained by linear interpolation on grid points
are used instead. (Cf. the \code{exact} argument of \code{\link[glmnet]{predict.glmnet}})}

\item{group_threshold}{If \code{TRUE}, (hard-)threshold coefficients so that
there is exactly one non-zero coefficient in each group.}

\item{...}{Additional arguments passed to \code{\link{exclusive_lasso}} if
\code{exact=TRUE} and ignored otherwise.}

\item{newx}{New data \eqn{X \in R^{m \times p}}{X} on which to make predictions. If not
supplied, predictions are made on trainng data.}

\item{type}{The type of "prediction" to return. If \code{type="link"}, returns
the linear predictor. If \code{type="response"}, returns the expected
value of the response. If \code{type="coefficients"}, returns the coefficients
used to calculate the linear predictor. (Cf. the \code{type} argument
of \code{\link[glmnet]{predict.glmnet}})}

\item{offset}{An offset term used in predictions. If not supplied, all offets are
taken to be zero. If the original fit was made with an offset, \code{offset} will
be required.}
}
\description{
Make predictions using the exclusive lasso. Similar to \code{\link[glmnet]{predict.glmnet}}.
\code{coef(...)} is a wrapper around \code{predict(..., type="coefficients")}.
}
\examples{
n <- 200
p <- 500
groups <- rep(1:10, times=50)
beta <- numeric(p);
beta[1:10] <- 3

X <- matrix(rnorm(n * p), ncol=p)
y <- X \%*\% beta + rnorm(n)

exfit <- exclusive_lasso(X, y, groups)
coef(exfit, lambda=1)
predict(exfit, lambda=1, newx = -X)
}
